Move statistics specifics from TranslationCommans to TranslationStatisticsCommand
